+++ django_request_formatter/forms.py
@@ -1,13 +1,13 @@
 import copy
 import json
 from collections import OrderedDict
-from typing import Dict
+from typing import Union
 
 import msgpack
-from django.core.exceptions import ValidationError
-from django.forms import MediaDefiningClass
+from django.core.exceptions import ValidationError, NON_FIELD_ERRORS
+from django.forms import MediaDefiningClass, Field
 
-from django_request_formatter.fields import Field
+from django_request_formatter.exceptions import RequestValidationError
 
 
 class DeclarativeFieldsMetaclass(MediaDefiningClass):
@@ -56,6 +56,7 @@ class BaseForm(object):
         self.fields = copy.deepcopy(getattr(self, 'base_fields'))
         self._errors = None
         self._dirty = []
+        self.cleaned_data = {}
 
         if isinstance(data, dict):
             for key in data.keys():
@@ -80,6 +81,9 @@ class BaseForm(object):
 
     @classmethod
     def create_from_request(cls, request):
+        """
+        :rtype: BaseForm
+        """
         if not request.body:
             return cls(None)
 
@@ -93,18 +97,73 @@ class BaseForm(object):
         return cls(data)
 
     @property
-    def errors(self) -> ValidationError:
+    def errors(self) -> dict:
+        if self._errors is None:
+            self.full_clean()
         return self._errors
 
-    @property
-    def payload(self) -> Dict:
-        result = {}
+    def is_valid(self) -> bool:
+        return not self.errors
+
+    def add_error(self, field: Union[str, None], error: Union[ValidationError, RequestValidationError]):
+        if isinstance(error, RequestValidationError):
+            self._errors[field] = error.errors
+            return
+
+        if hasattr(error, 'error_dict'):
+            if field is not None:
+                raise TypeError(
+                    "The argument `field` must be `None` when the `error` "
+                    "argument contains errors for multiple fields."
+                )
+            else:
+                error = error.error_dict
+        else:
+            error = {field or NON_FIELD_ERRORS: error.error_list}
+
+        for field, error_list in error.items():
+            if field not in self.errors:
+                if field != NON_FIELD_ERRORS and field not in self.fields:
+                    raise ValueError("'%s' has no field named '%s'." % (self.__class__.__name__, field))
+                if field == NON_FIELD_ERRORS:
+                    self._errors[field] = []
+                else:
+                    self._errors[field] = []
+            self._errors[field].extend(error_list)
+            if field in self.cleaned_data:
+                del self.cleaned_data[field]
+
+    def full_clean(self):
+        """
+        Clean all of self.data and populate self._errors and self.cleaned_data.
+        """
+        self._errors = {}
+        self.cleaned_data = {}
 
         for key, field in self.fields.items():
-            if key in self._dirty:
-                result[key] = field.to_python(self._data.get(key, None))
+            try:
+                self.cleaned_data[key] = field.clean(self._data.get(key, None))
+                if hasattr(self, f"clean_{key}"):
+                    self.cleaned_data[key] = getattr(self, f"clean_{key}")()
+            except (ValidationError, RequestValidationError) as e:
+                self.add_error(key, e)
 
-        return result
+        try:
+            cleaned_data = self.clean()
+        except ValidationError as e:
+            self.add_error(None, e)
+        else:
+            if cleaned_data is not None:
+                self.cleaned_data = cleaned_data
+
+    def clean(self):
+        """
+        Hook for doing any extra form-wide cleaning after Field.clean() has been
+        called on every field. Any ValidationError raised by this method will
+        not be associated with a particular field; it will have a special-case
+        association with the field named '__all__'.
+        """
+        return self.cleaned_data
 
     def fill(self, obj):
         """
@@ -115,41 +174,6 @@ class BaseForm(object):
         """
         for key, item in self._data.items():
             setattr(obj, key, item)
-
-    def is_valid(self, raise_exception: bool = True, **kwargs):
-        errors = {}
-
-        for key, field in self.fields.items():
-            field_errors = []
-            try:
-                field.validate(self._data.get(key, None))
-            except ValidationError as e:
-                field_errors.append(e)
-
-            if hasattr(self, f"validate_{key}"):
-                try:
-                    getattr(self, f"validate_{key}")(self._data.get(key, None), **kwargs)
-                except ValidationError as e:
-                    field_errors.append(e)
-
-            if field_errors:
-                errors[key] = field_errors
-
-        try:
-            self.validate(**kwargs)
-        except ValidationError as e:
-            errors = {**errors, **e.error_dict}
-
-        if errors:
-            self._errors = errors
-            if raise_exception:
-                raise ValidationError(errors)
-            return False
-
-        return True
-
-    def validate(self, **kwargs):
-        pass
 
 
 class Form(BaseForm, metaclass=DeclarativeFieldsMetaclass):