+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.5.0 : 16.12.2019
+
+- **Change**: Use native `django.form.fields` if possible
+- **Change**: Removed `kwargs` propagation from release `0.3.0`
+- **Change**: Changed syntax back to `django.forms` compatible (e.g. `form.validate_{key}()` -> `form.clean_{key}()`)
+- **Change**: `FieldList` raises `ValidationError` instead of `RuntimeException` if there is a type  in validation
+- **Change**: Use private properties for internal data in field objects
+- **Fixed**: `FieldList` returns values instead of `None`
+- **Fix**: Fixed validation in `DictionaryField`
+- **Maintenance**: Basic unit tests
+
 ## 0.4.3 : 29.11.2019
 
 - **Fix**: Fixed `Form` has no attribute `self._data`
@@ -16,7 +27,7 @@
 
 ## 0.3.0 : 11.11.2019
 
-- **Feature**: Propagate `kwargs` from `Form.is_valid()` to `Form.validate()` and `Form.validate_{key}()` methods.
+- **Feature**: Propagate `kwargs` from `Form.is_valid()` to `Form.validate()` and `Form.validate_{key}()` methods
 
 ## 0.2.1 : 4.11.2019
 
@@ -25,14 +36,14 @@
 ## 0.2.0 : 31.10.2019
 
 - **Change**: `Form.validate()` replaced by `Form.is_valid()`
-- **Feature**: `Form.validate()` is now used as a last step of form validation and it's aimed to be overwritten if 
-needed 
+- **Feature**: `Form.validate()` is now used as a last step of form validation and it's aimed to be overwritten if
+needed
 - **Note**: Unit tests initialization
 
 ## 0.1.6 : 24.10.2019
 
 - **Fix**: Non-required EnumField is now working
-- **Feature**: WIP: Initial method for filling objects `Form::fill()` 
+- **Feature**: WIP: Initial method for filling objects `Form::fill()`
 
 ## 0.1.5 : 23.10.2019